+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.15] - 2026-02-22
+
+### Added
+
+- Support for PH05 (Dyson Purifier Humidify+Cool De-Nox) - product type `358K`
+- Fix HP06 (Cryptomic) incorrectly mapped to product type `358K` - HP06 shares product type `527` with HP04
+
 ## [1.0.14] - 2026-02-21
 
 ### Fixed